# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Student Lost & Found Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  role: UserProfileRole;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  Lost: 'Lost',
  Found: 'Found',
} as const;

export interface Item {
  id: number;
  title: string;
  description: string;
  category: string;
  type: ItemType;
  location: string;
  date: string;
  /** @nullable */
  image?: string | null;
  contact: string;
  userId: number;
  createdAt: string;
  userName: string;
}

export type ItemInputType = typeof ItemInputType[keyof typeof ItemInputType];


export const ItemInputType = {
  Lost: 'Lost',
  Found: 'Found',
} as const;

export interface ItemInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  type: ItemInputType;
  location: string;
  date: string;
  contact: string;
  /** @nullable */
  image?: string | null;
}

export type ItemPatchType = typeof ItemPatchType[keyof typeof ItemPatchType];


export const ItemPatchType = {
  Lost: 'Lost',
  Found: 'Found',
} as const;

export interface ItemPatch {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  type?: ItemPatchType;
  location?: string;
  date?: string;
  contact?: string;
  /** @nullable */
  image?: string | null;
}

export interface DashboardData {
  items: Item[];
  totalLost: number;
  totalFound: number;
}

export interface StatsData {
  totalLost: number;
  totalFound: number;
  totalUsers: number;
  recentItems: Item[];
}

export interface AdminStatsData {
  totalUsers: number;
  totalItems: number;
  totalLost: number;
  totalFound: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export type ListItemsParams = {
/**
 * Search by item name
 */
search?: string;
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by Lost or Found
 */
type?: ListItemsType;
/**
 * Max items to return
 */
limit?: number;
/**
 * Pagination offset
 */
offset?: number;
};

export type ListItemsType = typeof ListItemsType[keyof typeof ListItemsType];


export const ListItemsType = {
  Lost: 'Lost',
  Found: 'Found',
} as const;