# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Lost & Found Portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: items
    description: Lost & Found items
  - name: dashboard
    description: User dashboard
  - name: stats
    description: Public statistics
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List all lost & found items with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by item name
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: type
          in: query
          schema:
            type: string
            enum: [Lost, Found]
          description: Filter by Lost or Found
        - name: limit
          in: query
          schema:
            type: number
          description: Max items to return
        - name: offset
          in: query
          schema:
            type: number
          description: Pagination offset
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"

    post:
      operationId: createItem
      tags: [items]
      summary: Post a new lost or found item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get a specific item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateItem
      tags: [items]
      summary: Update an item (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemPatch"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item (owner or admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get the current user's dashboard (their items + stats)
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getPublicStats
      tags: [stats]
      summary: Get public site statistics for the landing page
      responses:
        "200":
          description: Public stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsData"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/items:
    get:
      operationId: listAdminItems
      tags: [admin]
      summary: List all items (admin only)
      responses:
        "200":
          description: All items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/items/{id}:
    delete:
      operationId: deleteAdminItem
      tags: [admin]
      summary: Delete any item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin-level statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatsData"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth schemas
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"

    UserProfile:
      type: object
      required: [id, name, email, createdAt, role]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [student, admin]

    # Item schemas
    Item:
      type: object
      required: [id, title, description, category, type, location, date, contact, userId, createdAt, userName]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [Lost, Found]
        location:
          type: string
        date:
          type: string
        image:
          type: ["string", "null"]
        contact:
          type: string
        userId:
          type: number
        createdAt:
          type: string
          format: date-time
        userName:
          type: string

    ItemInput:
      type: object
      required: [title, description, category, type, location, date, contact]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [Lost, Found]
        location:
          type: string
        date:
          type: string
        contact:
          type: string
        image:
          type: ["string", "null"]

    ItemPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [Lost, Found]
        location:
          type: string
        date:
          type: string
        contact:
          type: string
        image:
          type: ["string", "null"]

    # Dashboard schemas
    DashboardData:
      type: object
      required: [items, totalLost, totalFound]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        totalLost:
          type: number
        totalFound:
          type: number

    # Stats schemas
    StatsData:
      type: object
      required: [totalLost, totalFound, totalUsers, recentItems]
      properties:
        totalLost:
          type: number
        totalFound:
          type: number
        totalUsers:
          type: number
        recentItems:
          type: array
          items:
            $ref: "#/components/schemas/Item"

    AdminStatsData:
      type: object
      required: [totalUsers, totalItems, totalLost, totalFound]
      properties:
        totalUsers:
          type: number
        totalItems:
          type: number
        totalLost:
          type: number
        totalFound:
          type: number

    AdminUser:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time