# lib/api-zod/src/generated/types/adminStatsData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Student Lost & Found Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStatsData {
  totalUsers: number;
  totalItems: number;
  totalLost: number;
  totalFound: number;
}