# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Student Lost & Found Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemType } from './itemType';

export interface Item {
  id: number;
  title: string;
  description: string;
  category: string;
  type: ItemType;
  location: string;
  date: string;
  /** @nullable */
  image?: string | null;
  contact: string;
  userId: number;
  createdAt: Date;
  userName: string;
}