# lib/api-zod/src/generated/types/itemPatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Student Lost & Found Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemPatchType } from './itemPatchType';

export interface ItemPatch {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  type?: ItemPatchType;
  location?: string;
  date?: string;
  contact?: string;
  /** @nullable */
  image?: string | null;
}