# lib/api-zod/src/generated/types/statsData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Student Lost & Found Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Item } from './item';

export interface StatsData {
  totalLost: number;
  totalFound: number;
  totalUsers: number;
  recentItems: Item[];
}